backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
bcrypt==3.2.2
python-jose[cryptography]
pydantic[email]
python-dotenv
